+++ modules/frames/src/+scenario/+frames/Contents.m
@@ -2,5 +2,6 @@
 %
 %   convertGeodeticToEarthFixed - Convert WGS84 LLA to ECEF position.
 %   convertEarthFixedToGeodetic - Convert ECEF position to WGS84 LLA.
+%   calculateEarthFixedToInertialRotation - Calculate ECEF-to-ECI rotations.
 %   convertEarthFixedToInertial - Convert ECEF positions to ECI.
 %   convertInertialToEarthFixed - Convert ECI positions to ECEF.